[Spec 55][Phase: fix] refactor: Trim heavyweight diagnostics from canonical suite (FR2)
Phase-3 iter1 consult: Claude APPROVE; Gemini + Codex REQUEST_CHANGES on the same
FR2 point — the canonical tests/e2e/graph-handle.ts still carried heavyweight
phase-1/2 diagnostics. Spec FR2 keeps committed only (a) fix dependencies and
(b) the node-occupancy / background-point helper; it names verbose per-event
pointer logs, controls.enabled sampling traces, and dumped-on-failure counters as
evidence-only.

Relocate (not delete — keeps the diagnostic re-runnable, per plan FR2 state) the
heavyweight helpers out of the canonical suite:
- NEW tests/diagnostics/55-drag/drag-probe.ts: PointerLog/PointerEventRecord/
  ControlsSample types, installDragProbe() (capture-phase pointer counters +
  fast controls sampler via its own addInitScript + a minimal duplicated
  findHandle), and sampleControls/readPointerLog/resetPointerLog wrappers.
- tests/e2e/graph-handle.ts trimmed to the H1 fix dependency only: kept
  NodeOccupancy/__graphNodeOccupancyAtPoint/nodeOccupancyAtPoint and
  pickBackgroundDragPoint/BackgroundDragPoint; removed the 3 heavyweight types,
  4 Window decls, the pointer-counter install block, __graphControlsSample, and
  3 wrappers.
- drag-diagnostic.spec.ts: split imports (keepable from graph-handle, heavyweight
  from ./drag-probe) + installDragProbe(page) before openGraphPage.

Verified: typecheck + eslint clean; canonical --list unchanged (0 diagnostic
tests, 18 matrix tests); canonical :224 fix green 4/4 both engines; diagnostic
still functions post-move (freshly reproduced H1 on firefox with valid
occupancy/counters/controls data). Fix itself unchanged.

Includes iter1 consult outputs + rebuttal.
